requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.3
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pymupdf>=1.23.21
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4
numpy>=1.26.3
pandas>=2.2.0
python-dotenv>=1.0.1
httpx>=0.26.0
bcrypt==3.2.0
email-validator
